backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
httpx==0.28.*
python-dotenv==1.1.*
langchain
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
langchain-openai
langchain-chroma
chromadb
rank_bm25
pypdf